The return type is an int. Casting it to a size_t before checking if it's bigger than zero lead to a true condition ... always ...
This commit was SVN r8114.

+++ orte/mca/rmaps/round_robin/rmaps_rr_component.c
@@ -68,7 +68,7 @@ orte_rmaps_round_robin_component_t mca_rmaps_round_robin_component = {
   */
 static int orte_rmaps_round_robin_open(void)
 {
-    size_t id;
+    int id;
 
     mca_base_param_reg_int(&mca_rmaps_round_robin_component.super.rmaps_version, "debug",
                            "Toggle debug output for Round Robin RMAPS component",
